Moved the stuff students wrote into the Hoops bot
