+++ Hoops/src/main/java/org/firstinspires/ftc/hoops/Setup.java
@@ -8,31 +8,25 @@ public class Setup {
     public static class Connected {
 
         public static boolean DRIVEBASE = true;
-        public static boolean TESTSUBSYSTEM = false;
-        public static boolean MOTOR = false;
-        public static boolean SERVO = false;
-        public static boolean DISTANCE_SENSOR = false;
-        public static boolean COLOR_SENSOR = false;
-        public static boolean FLYWHEEL = false;
-        public static boolean WEBCAM = false;
         public static boolean EXTERNAL_IMU = true;
+        public static boolean LAUNCHER = true;
+        public static boolean INTAKE = true;
     }
 
     @Config
     public static class HardwareNames {
 
-        public static String MOTOR = "motor";
         public static String FLMOTOR = "fl";
         public static String FRMOTOR = "fr";
         public static String RLMOTOR = "rl";
         public static String RRMOTOR = "rr";
-        public static String FLYWHEELMOTOR = "fly";
-        public static String SERVO = "s";
         public static String IMU = "imu";
         public static String EXTERNAL_IMU = "adafruit-imu";
-        public static String DISTANCE = "d";
-        public static String COLOR = "c";
-        public static String CAMERA = "camera";
+
+        public static String TOP_LAUNCH = "top";
+        public static String BOTTOM_LAUNCH = "bottom";
+
+        public static String INTAKE = "slurp";
     }
 
     @Config
