+++ Hoops/src/main/java/org/firstinspires/ftc/hoops/subsystems/DrivebaseSubsystem.java
@@ -154,7 +154,6 @@ public class DrivebaseSubsystem
         rl2 = rlMotor;
         rr2 = rrMotor;
         curHeading = imu.getHeading();
-        Setup.HardwareNames.COLOR = imu.getClass().toString();
     }
 
     @Override
@@ -167,7 +166,6 @@ public class DrivebaseSubsystem
             " : " +
             (poseVelocity != null ? poseVelocity.toString() : "nullv");
             curHeading = this.gyro.getHeading();
-            Setup.HardwareNames.FLYWHEELMOTOR = this.gyro.getClass().toString();
         }
     }
