+++ Hoops/src/main/java/org/firstinspires/ftc/hoops/subsystems/IntakeSubsystem.java
@@ -1,9 +1,9 @@
-package org.firstinspires.ftc.twenty403.subsystems;
+package org.firstinspires.ftc.hoops.subsystems;
 
 import com.qualcomm.robotcore.hardware.DcMotorEx;
 import com.qualcomm.robotcore.hardware.DistanceSensor;
 import com.technototes.library.hardware.motor.EncodedMotor;
-import org.firstinspires.ftc.twenty403.Hardware;
+import org.firstinspires.ftc.hoops.Hardware;
 
 public class IntakeSubsystem {
 
@@ -13,7 +13,7 @@ public class IntakeSubsystem {
     public static double THE_VELOCITY = 0.5;
 
     public IntakeSubsystem(Hardware h) {
-        slurp = h.armR;
+        slurp = h.slurp;
         slurp.coast();
 
         //TODO: replace null with something
